requirements.txt
Adyen==4.0.0
aiohttp==3.7.4.post0
aiopg==1.3.1
amqp==5.3.1
aniso8601==7.0.0
anyio==2.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.1.1
asgiref==3.2.10
asn1crypto==1.4.0
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
authorizenet==1.1.4
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
billiard==4.2.1
binaryornot==0.4.4
bleach==3.3.0
boto3==1.18.21
botocore==1.21.21
bottle==0.12.19
braintree==4.11.0
bs4==0.0.1
cachetools==4.2.2
cairocffi==1.2.0
CairoSVG==2.5.2
capstone==4.0.2
celery==5.4.0
certifi==2020.6.20
cffi==1.14.6
channels==3.0.3
channels-postgres==0.0.3
chardet==3.0.4
chargily_pay==2.1.3
charset-normalizer==2.0.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
constantly==15.1.0
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect2==0.4.1
daphne==3.0.2
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.0.7
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.5
django-allauth==0.44.0
django-analytical==3.1.0
django-appconf==1.0.4
django-autoslug==1.9.9
django-bootstrap4==2.3.1
django-braces==1.14.0
django-cache-url==3.2.3
django-channels==0.7.0
django-channels-graphql-ws==0.8.0
django-chartjs==2.3.0
django-cities-light==3.10.0
django-ckeditor==6.7.0
django-colorfield==0.11.0
django-colorful==1.3
django-cors-headers==3.7.0
django-countries==7.2.1
django-crispy-forms==1.10.0
django-currencies==0.10.1
django-debug-toolbar==3.2
django-disqus==0.5
django-environ==0.9.0
django-extensions==3.1.0
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-grappelli==4.0.1
django-haystack==3.0
django-imagekit==5.0.0
django-import-export==0.5.0
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-jsonfield-backport==1.0.3
django-markdown==0.8.4
django-measurement==3.2.3
django-model-utils==4.0.0
django-mptt==0.12.0
django-nocaptcha-recaptcha==0.0.19
django-notifications==0.1.dev0
django-pagedown==2.2.0
django-pandas==0.6.4
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-prices==2.2.0
django-prices-openexchangerates==1.1.0
django-prices-vatlayer==1.1.0
django-private-chat==0.3.0
django-redis==5.0.0
django-robots==6.1
django-rocket-engine==0.0.11
django-rosetta==0.9.7
django-simditor==0.0.15
django-social-share==2.2.1
django-storages==1.11.1
django-suit==0.2.21
django-templated-mail==1.1.1
django-tenants==3.3.2
django-tinymce==4.1.0
django-versatileimagefield==2.1
django-widget-tweaks==1.4.8
django-wysiwyg-redactor==0.4.9.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
draftjs-sanitizer==1.0.0
enmerkar==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
facebook-sdk==3.1.0
fade==0.0.9
Faker==8.11.0
filelock==3.0.12
freezegun==1.1.0
generic==1.0.0
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-core==1.7.2
google-cloud-pubsub==2.7.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-i18n-address==2.5.0
google-measurement-protocol==1.1.0
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
googletrans==4.0.0rc1
graphene==2.1.9
graphene-django==2.13.0
graphene-federation==0.1.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
html-to-draftjs==1.0.1
html2text==2020.1.16
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.8
ignorant==1.2
incremental==21.3.0
intervaltree==3.1.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
itypes==1.2.0
jaeger-client==4.6.0
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
json5==0.9.5
jsonfield==2.1.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-server==1.3.0
jupyterlab==3.0.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.2.0
kombu==5.4.2
lazy-object-proxy==1.4.3
libcst==0.3.20
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
markdown2==2.4.0
MarkupSafe==1.1.1
maxminddb==2.0.3
mccabe==0.6.1
measurement==3.2.0
micawber==0.5.4
mistune==0.8.4
mock==2.0.0
mpmath==1.2.1
msgpack==1.0.2
multidict==5.1.0
mypy-extensions==0.4.3
mysql==0.0.3
mysqlclient==2.0.3
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
notifications==0.3.2
numpy==1.21.1
oauthlib==3.1.1
olefile==0.46
openpyxl==3.0.7
opentracing==2.4.0
outcome==1.2.0
packaging==21.0
pafy==0.5.5
pandas==1.3.1
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.1
pbr==1.10.0
pdfkit==1.0.0
petl==1.7.4
phonenumbers==8.13.50
phonenumberslite==8.12.29
pickleshare==0.7.5
pilkit==3.0
Pillow==10.0.1
pipenv==2020.11.15
polib==1.1.1
poyo==0.5.0
prices==1.1.0
progressbar2==4.2.0
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.19
proto-plus==1.19.0
protobuf==3.17.3
psutil==5.8.0
psycopg2==2.8.6
psycopg2-binary==2.9.1
pwntools==4.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybars3==0.9.7
pycairo==1.20.1
pycodestyle==2.7.0
pycparser==2.20
pyelftools==0.27
Pygments==2.7.4
PyHamcrest==2.0.2
PyJWT==2.1.0
pylint==2.6.0
PyMeta3==0.5.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyphen==0.11.0
pyrsistent==0.17.3
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.1
python-dotenv==0.19.0
python-gettext==4.0
python-http-client==3.3.2
python-json-logger==2.0.2
python-magic==0.4.24
python-magic-bin==0.4.14
python-memcached==1.59
python-slugify==4.0.1
python-social-auth==0.3.6
python-utils==3.8.1
python3-openid==3.1.0
pytimeparse==1.1.8
pytube==10.4.1
pytz==2024.2
pywin32==300
pywinpty==0.5.7
PyXB==1.2.5
PyYAML==5.4.1
pyzmq==22.0.2
raven==6.10.0
razorpay==1.2.0
redis==3.5.3
reportlab==4.0.6
requests==2.31.0
requests-futures==1.0.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
ROPGadget==6.5
rsa==4.7.2
Rx==1.6.1
s3transfer==0.5.0
selenium==3.141.0
Send2Trash==1.5.0
sendgrid==6.8.0
sentry-sdk==1.3.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.15.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.10.0
sortedcontainers==2.3.0
soupsieve==2.0.1
speedtest-cli==2.1.2
sqlparse==0.3.1
starkbank-ecdsa==1.1.1
stem==1.8.1
stripe==2.60.0
sympy==1.8
tablib==0.11.2
termcolor==2.1.0
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadloop==1.0.2
thrift==0.13.0
tinycss2==1.1.0
toml==0.10.2
tornado==6.1
torrequest==0.1.0
tqdm==4.57.0
traitlets==5.0.5
trio==0.21.0
trio-websocket==0.9.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typed-ast==1.4.2
typing==3.7.4.3
typing-inspect==0.7.1
typing_extensions==4.12.2
tzdata==2023.3
undetected-chromedriver==3.1.5.post4
unicorn==1.0.2rc3
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.25.11
uvicorn==0.13.4
vine==5.1.0
virtualenv==20.4.0
virtualenv-clone==0.5.4
watchgod==0.7
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
WeRoBot==1.13.1
whitenoise==5.2.0
Whoosh==2.7.4
wrapt==1.12.1
wsproto==1.1.0
xmltodict==0.12.0
yarl==1.6.3
youtube-dl==2020.12.14
zope.interface==5.4.0
